requirements.txt.txt
pandas
numpy
scikit-surprise
scikit-learn
streamlit
